backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.2
billiard==4.2.4
celery==5.6.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cmake==4.2.1
cryptography==46.0.5
dlib==20.0.0
ecdsa==0.19.1
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
fastapi==0.129.2
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
h11==0.16.0
httplib2==0.31.2
idna==3.11
kombu==5.6.2
mysql-connector-python==9.6.0
numpy==2.4.2
oauthlib==3.3.1
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
packaging==26.2
pandas==3.0.1
passlib==1.7.4
pillow==12.1.1
prompt_toolkit==3.0.52
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.11.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
redis==7.4.0
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==69.5.1
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
